requirements.txt
fastapi~=0.115.0
pydantic~=2.9.2
passlib~=1.7.4
pymongo~=4.9.1
python-dotenv~=1.0.1
requests~=2.32.3
certifi~=2024.8.30